+++ src/scripts.js
@@ -152,32 +152,32 @@ function initXopatScripts() {
                 let zoom = null,
                     speed = 0.3;
                 switch (e.key) {
-                case "Down": // IE/Edge specific value
-                case "ArrowDown":
-                    adjustBounds(0, speed);
-                    break;
-                case "Up": // IE/Edge specific value
-                case "ArrowUp":
-                    adjustBounds(0, -speed);
-                    break;
-                case "Left": // IE/Edge specific value
-                case "ArrowLeft":
-                    adjustBounds(-speed, 0);
-                    break;
-                case "Right": // IE/Edge specific value
-                case "ArrowRight":
-                    adjustBounds(speed, 0);
-                    break;
-                case "+":
-                    zoom = VIEWER.viewport.getZoom();
-                    VIEWER.viewport.zoomTo(zoom + zoom * speed * 3);
-                    return;
-                case "-":
-                    zoom = VIEWER.viewport.getZoom();
-                    VIEWER.viewport.zoomTo(zoom - zoom * speed * 2);
-                    return;
-                default:
-                    return; // Quit when this doesn't handle the key event.
+                    case "Down": // IE/Edge specific value
+                    case "ArrowDown":
+                        adjustBounds(0, speed);
+                        break;
+                    case "Up": // IE/Edge specific value
+                    case "ArrowUp":
+                        adjustBounds(0, -speed);
+                        break;
+                    case "Left": // IE/Edge specific value
+                    case "ArrowLeft":
+                        adjustBounds(-speed, 0);
+                        break;
+                    case "Right": // IE/Edge specific value
+                    case "ArrowRight":
+                        adjustBounds(speed, 0);
+                        break;
+                    case "+":
+                        zoom = VIEWER.viewport.getZoom();
+                        VIEWER.viewport.zoomTo(zoom + zoom * speed * 3);
+                        return;
+                    case "-":
+                        zoom = VIEWER.viewport.getZoom();
+                        VIEWER.viewport.zoomTo(zoom - zoom * speed * 2);
+                        return;
+                    default:
+                        return; // Quit when this doesn't handle the key event.
                 }
             }
 
@@ -261,10 +261,14 @@ function initXopatScripts() {
             point: VIEWER.viewport.getCenter()
         };
 
+        //delete unnecessary data
+        const data = {...APPLICATION_CONTEXT.config};
+        delete data.defaultParams;
+
         //by default ommit underscore
         let app = APPLICATION_CONTEXT.layersAvailable && window.WebGLModule
-            ? JSON.stringify(APPLICATION_CONTEXT.config, WebGLModule.jsonReplacer)
-            : JSON.stringify(APPLICATION_CONTEXT.config, (key, value) => key.startsWith("_") ? undefined : value);
+            ? JSON.stringify(data, WebGLModule.jsonReplacer)
+            : JSON.stringify(data, (key, value) => key.startsWith("_") ? undefined : value);
         APPLICATION_CONTEXT.config.params.viewport = oldViewport;
         APPLICATION_CONTEXT.config.params.bypassCookies = bypass;
         APPLICATION_CONTEXT.config.params.bypassCacheLoadTime = false;
@@ -429,10 +433,9 @@ ${await UTILITIES.getForm()}
 
     /**
      * Refresh current page with all plugins and their data if export API used
-     * @param formInputHtml additional HTML to add to the refresh FORM
      * @param includedPluginsList of ID's of plugins to include, inludes current active if not specified
      */
-    window.UTILITIES.refreshPage = async function(formInputHtml="", includedPluginsList=undefined) {
+    window.UTILITIES.refreshPage = async function(includedPluginsList=undefined) {
         if (APPLICATION_CONTEXT.__cache.dirty) {
             Dialogs.show($.t('messages.warnPageReload', {
                 onExport: "UTILITIES.export();",
@@ -441,12 +444,8 @@ ${await UTILITIES.getForm()}
             return;
         }
 
-        // if (window.removeEventListener) {
-        //     window.removeEventListener('beforeunload', preventDirtyClose, true);
-        // } else if (window.detachEvent) {
-        //     window.detachEvent('onbeforeunload', preventDirtyClose);
-        // }
-        $(document.body).append(await UTILITIES.getForm(formInputHtml, includedPluginsList, true));
+        UTILITIES.storePageState(includedPluginsList);
+        window.location.replace(APPLICATION_CONTEXT.url);
     };
 
     /**